requirements.txt
django
djangorestframework
django-storages
python-decouple
boto3